order-api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "order-api"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "boto3==1.34.162",
    "pydantic==2.8.2",
    "pydantic-settings==2.4.0",
    "slowapi==0.1.9",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.9",
]

[project.optional-dependencies]
test = [
    "pytest==8.3.2",
    "httpx==0.27.0",
    "pytest-asyncio==0.23.8",
    "pyjwt==2.10.1",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
asyncio_mode = "auto"